Moved ContextMenuTests, IssuePanelTests and KeyboardShortcutsTest to unstable.
