+++ src/test/java/unstable/ContextMenuTests.java
@@ -1,8 +1,10 @@
-package guitests;
+package unstable;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
+
+import guitests.UITest;
 import javafx.application.Platform;
 import javafx.scene.control.ContextMenu;
 import javafx.scene.control.MenuItem;
